0610 parsing & p,s,r,rr

+++ includes/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dim <dim@student.42seoul.kr>               +#+  +:+       +#+        */
+/*   By: dim <dim@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/25 02:33:16 by dim               #+#    #+#             */
-/*   Updated: 2021/06/09 17:13:06 by dim              ###   ########.fr       */
+/*   Updated: 2021/06/10 20:15:03 by dim              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ typedef struct		s_list
 }					t_list;
 
 int					ft_atoi(const char *str);
-int					ft_atol(const char *str);
+long long			ft_atol(const char *str);
 void				ft_bzero(void *s, size_t n);
 void				*ft_calloc(size_t n, size_t size);
 int					ft_isalnum(int c);
